Move indentation tests into tests folder
